Start consolidating ActionBar to TS
