+++ app/components/listing/index.tsx
@@ -1,5 +1,5 @@
-export { default as ActionSidebar } from './ActionSidebar';
-export { default as MobileActionBar } from './MobileActionBar';
+export * from './ActionSidebar';
+export * from './ActionSidebarMobile';
 export * from './ListingTitleLink';
 export * from './MapOfLocations';
 export * from './MOHCDBadge';
